Add simpler configuration syntax and language specific block markers.
Simplifications:
  - allow configurations without `re2c` prefix (for languages other than C/C++
    `re2c` prefix is illogical, as the program has a different name, e.g.
    `re2ocaml` for OCaml, etc.)
  - allow dropping `define:` part
  - allow dropping `variable:` part

So, for example, former `re2c:define:YYFILL` is now `:YYFILL`.

Language-specific block markers are of the form `/*!re2<x>`, where `<x>` is a
word that should match the name of the program (e.g. `ocaml` for re2ocaml), or
else `<x>` can be `c` (historically `/*!re2c` marker was allowed for languages
other than C/C++, we should allow it for backwards compatibility).