Playground: use default monospace font, set editor theme.
Although Iosevka is a nice font, it may be a bit hard to read as it's very
narrow. It is better to use default system or browser monospace font, which
should be familiar to the user (and this is what we do for the examples on
re2c.org anyway).