Generate correct escape sequences for each language.
Previously C escape sequences were used for all languages, which resulted in
incorrect escapes generated for some languages, e.g. `\f` in OCaml is not a
valid escape, while in C it is.

This commit adds syntax configuration `special_escapes` to all syntax files.
Its value is a string of all special characters that require escaping.

This is a fix for #516 "re2ocaml can generate invalid escape sequences".