Add configuration `re2c:on-default`.
The new configuration controls default rule behaviour.

Option `re2c:on-default = match-code-point` is the default one, it keeps the
current behaviour: default rule matches a single code point.

Option `re2c:on-default = error` introduces new a behaviour: default rule
matches the "error location", that is, it consumes all code points up to the
fallback point (there's no YYRESTORE operation). In addition, the postion of
the previous match is backed up (there's an YYBACKUP operation in the start
state). In terms of simple API, YYCURSOR is set to the fallback location, and
YYMARKER is set the the previous match location.

This addresses #539: "feature request: index of illegal character".