Revert "Add syntax variable `attrs` for `code:fndecl` and `code:fndef` configurations."
This reverts commit 85d9384b2f95b3de0464b82a53680c8956dfc01c.

Reason for revert: the main motivation for this change was to support Swift
`throw` annotations. However, it turns out that attributes are unfit for the
purpose, since functions that throw also require slightly different call
expressions (they need to be prefixed with `try`). Although attributes may be
useful, let's remove them until we find a real-world use case.

For details, see the full conversation:
https://github.com/skvadrik/re2c/pull/538#discussion_r2008079095