Zig: support `yyfn:throw` configuration.
This is useful for --recursive-functions code model, as it's necessary to use
different syntax for function calls with an additional `try` keyword before the
call expression (which would be impossible to do before this change).

The test/codegen/zig/recursive_functions/basic.re test has been updated to
demonstrate the new functionality.